Minor Improvements
Add Witch alias to Mage
Make damage a double

+++ src/com/gmail/scyntrus/fmob/FactionMobs.java
@@ -166,11 +166,11 @@ public class FactionMobs extends JavaPlugin {
 		Titan.maxHp = (float) config.getDouble("Titan.maxHp", Titan.maxHp);
 		if (Titan.maxHp<1) Titan.maxHp = 1;
 		
-		Archer.damage = config.getInt("Archer.damage", Archer.damage);
+		Archer.damage = config.getDouble("Archer.damage", Archer.damage);
 		if (Archer.damage<0) Archer.damage = 0;
-		Swordsman.damage = config.getInt("Swordsman.damage", Swordsman.damage);
+		Swordsman.damage = config.getDouble("Swordsman.damage", Swordsman.damage);
 		if (Swordsman.damage<0) Swordsman.damage = 0;
-		Titan.damage = config.getInt("Titan.damage", Titan.damage);
+		Titan.damage = config.getDouble("Titan.damage", Titan.damage);
 		if (Titan.damage<0) Titan.damage = 0;
 		
 		Archer.enabled = config.getBoolean("Archer.enabled", Archer.enabled);
