+++ src/com/gmail/scyntrus/fmob/mobs/Swordsman.java
@@ -52,7 +52,7 @@ public class Swordsman extends EntitySkeleton implements FactionMob {
 	public static double powerCost = 0;
 	public static double moneyCost = 0;
 	public static double range = 16;
-	public static int damage = 0;
+	public static double damage = 0;
 	public static int drops = 0;
 	private int retargetTime = 0;
 	private double moveSpeed;
